+++ storm-core/test/clj/backtype/storm/cluster_test.clj
@@ -25,7 +25,7 @@
   (:require [conjure.core])
   (:use [conjure core])
   (:use [clojure test])
-  (:use [backtype.storm cluster config util testing]))
+  (:use [backtype.storm cluster config util testing thrift log]))
 
 (defn mk-config [zk-port]
   (merge (read-storm-config)
@@ -168,10 +168,10 @@
 (deftest test-storm-cluster-state-basics
   (with-inprocess-zookeeper zk-port
     (let [state (mk-storm-state zk-port)
-          assignment1 (Assignment. "/aaa" {} {1 [2 2002 1]} {})
-          assignment2 (Assignment. "/aaa" {} {1 [2 2002]} {})
-          base1 (StormBase. "/tmp/storm1" 1 {:type :active} 2 {} "")
-          base2 (StormBase. "/tmp/storm2" 2 {:type :active} 2 {} "")]
+          assignment1 (Assignment. "/aaa" {} {[1] ["1" 1001 1]} {})
+          assignment2 (Assignment. "/aaa" {} {[2] ["2" 2002]} {})
+          base1 (StormBase. "/tmp/storm1" 1 {:type :active} 2 {} "" nil nil)
+          base2 (StormBase. "/tmp/storm2" 2 {:type :active} 2 {} "" nil nil)]
       (is (= [] (.assignments state nil)))
       (.set-assignment! state "storm1" assignment1)
       (is (= assignment1 (.assignment-info state "storm1" nil)))
@@ -242,20 +242,21 @@
 (deftest test-supervisor-state
   (with-inprocess-zookeeper zk-port
     (let [state1 (mk-storm-state zk-port)
-          state2 (mk-storm-state zk-port)]
+          state2 (mk-storm-state zk-port)
+          supervisor-info1 (SupervisorInfo. 10 "hostname-1" "id1" [1 2] [] {} 1000 )
+          supervisor-info2 (SupervisorInfo. 10 "hostname-2" "id2" [1 2] [] {} 1000 )
+          ]
       (is (= [] (.supervisors state1 nil)))
-      (.supervisor-heartbeat! state2 "2" {:a 1})
-      (.supervisor-heartbeat! state1 "1" {})
-      (is (= {:a 1} (.supervisor-info state1 "2")))
-      (is (= {} (.supervisor-info state1 "1")))
+      (.supervisor-heartbeat! state2 "2" supervisor-info2)
+      (.supervisor-heartbeat! state1 "1" supervisor-info1)
+      (is (= supervisor-info2 (.supervisor-info state1 "2")))
+      (is (= supervisor-info1 (.supervisor-info state1 "1")))
       (is (= #{"1" "2"} (set (.supervisors state1 nil))))
       (is (= #{"1" "2"} (set (.supervisors state2 nil))))
       (.disconnect state2)
       (is (= #{"1"} (set (.supervisors state1 nil))))
       (.disconnect state1)
       )))
-
-
 
 (deftest test-cluster-authentication
   (with-inprocess-zookeeper zk-port